requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
gunicorn==25.1.0
python-dotenv==1.1.1
Pillow==11.2.1
requests==2.32.5
flutterwave-python==2.0.0
django-allauth==65.10.0
django-filter==25.1